Refactor dashboard and relating pages
