+++ client/scripts/components/dashboard.jsx
@@ -3,7 +3,6 @@
 // dependencies
 var React = require('react/addons');
 var ElectionActions = require('../actions/ElectionActions');
-var Link = require('react-router').Link;
 var _ = require('underscore');
 
 // store and actions
@@ -25,18 +24,12 @@ var IndexComponent = React.createClass({
     router: React.PropTypes.func
   },
 
-  goVote: function () {
-    var election = this.state.voteText;
-    this.context.router.transitionTo('/vote/' + election);
-  },
-
-  whoWon: function () {
-    var election = this.state.resultsText;
-    this.context.router.transitionTo('/elections/' + election);
-  },
-
   createElection: function () {
     ElectionActions.createElection()
+  },
+
+  createGroup: function(){
+    this.context.router.transitionTo('/dashboard/group/create');
   },
 
   getInitialState: function () {
@@ -75,14 +68,10 @@ var IndexComponent = React.createClass({
           <ul>{elections}</ul>
           <button onClick={this.createElection}>Create a new Election</button>
         </div>
-        <h2>Vote in an Election:</h2>
-          <input type='text' placeholder='Enter election ID' valueLink={this.linkState('voteText')}/>
-          <button onClick={this.goVote}>Vote!</button>
-        <h2>View Results</h2>
-          <input type='text' placeholder='Enter election ID' valueLink={this.linkState('resultsText')} />
-          <button onClick={this.whoWon}>Who won?</button>
+        <button onClick={this.createGroup}>Create a new Group</button>
         <h5><a href='/docs/api/index.html'>API Reference</a></h5>
         <h5><a href='/docs/styleguide/index.html'>Style Guide</a></h5>
+
       </div>
     );
   },
